src/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
Authlib==1.6.1
certifi==2025.8.3
cffi==1.17.1
click==8.2.1
cryptography==45.0.5
fastapi==0.116.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
sniffio==1.3.1
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
